upgraded crypto-js from 3 to 4 (vulnerability scan)
Overview
crypto-js is a library of crypto standards.

Affected versions of this package are vulnerable to Insecure Randomness. The secureRandom() method is supposed to return a cryptographically strong pseudo-random data string, but it is biased to certain digits. An attacker could be able to guess the created digits.

Remediation
Upgrade crypto-js to version 4.0.0, 3.2.1 or higher.